Move Status related components into a StatusPanel component (so updating status only affects rendres of related components).
